fix: complete cli_commands split and remove dead code from ups_rs_client
- Add cli_commands.py with all command handlers extracted from cli.py
- Fix _event_handler in cli_commands to also print procedure step state
  (00741000), resolving the failing test_event_handler assertion
- Remove dead _event_handler wrapper and main() wrapper from ups_rs_client.py
  (backward-compat shims that pointed to non-existent cli._event_handler)
- Remove __main__ block from ups_rs_client.py (CLI now lives in cli.py)
- Fix test_cli_validation.py import: main now imported from cli, not ups_rs_client

All 148 tests pass. Zero linting violations. Zero mypy errors.
ups_rs_client.py: 1453 → 1428 lines.
